환율 크롤링 수정

+++ server/src/main/java/com/ict/extravel/domain/pointexchange/service/KakaoPayService.java
@@ -87,7 +87,7 @@ public class KakaoPayService {
         PointCharge pointCharge = PointCharge.builder()
                 .tid(payReadyResDto.getTid())
                 .member(member)
-                .amount(Integer.parseInt(payInfoDto.getPrice()))
+                .amount(BigDecimal.valueOf(Integer.parseInt(payInfoDto.getPrice())))
                 .plusPoint(plusPoint)
                 .createdAt(createdAt)
                 .approvedAt(null)
